refactor(Card): 카드 모델로 이동
