+++ src/main/java/dto/PlayerResult.java
@@ -1,6 +1,7 @@
 package dto;
 
 import java.util.List;
+import model.Card;
 
 public record PlayerResult(String name, List<Card> hand, int score) {
     public PlayerResult {
