+++ src/main/java/model/BlackJackDeck.java
@@ -3,7 +3,6 @@ package model;
 import exception.GameException;
 import java.util.ArrayList;
 import java.util.List;
-import dto.Card;
 
 public class BlackJackDeck {
     private final List<Card> cards;
