+++ src/main/java/model/CardFactory.java
@@ -1,6 +1,5 @@
 package model;
 
-import dto.Card;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
