+++ src/main/java/model/ParticipantHand.java
@@ -1,7 +1,6 @@
 package model;
 
 import exception.GameException;
-import dto.Card;
 import java.util.ArrayList;
 import java.util.List;
