+++ src/main/java/service/BlackJackService.java
@@ -1,6 +1,6 @@
 package service;
 
-import dto.Card;
+import model.Card;
 import dto.ParticipantWinning;
 import dto.PlayerProfit;
 import java.math.BigDecimal;
