+++ src/test/java/model/ParticipantHandTest.java
@@ -2,7 +2,6 @@ package model;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import dto.Card;
 import java.util.List;
 import org.junit.jupiter.api.Test;
